fix: corrigindo a possibilidade de deletar um usuario nao existente

+++ src/modules/users/useCases/deleteUser/deleteUserUseCase.ts
@@ -7,6 +7,12 @@ interface IRequest {
 class DeleteUserUseCase {
     constructor(private usersRepository: UsersRepository){}
     execute({user_id}:IRequest){
+        const existingUser = this.usersRepository.findById(user_id);
+
+        if(!existingUser){
+            throw new Error("Usuário não encontrado!");
+        }
+        
         const user = this.usersRepository.deleteUser(user_id);
         return user;
     }
